javassist 满标复审奖励完善

+++ fanyin-service/src/main/java/com/fanyin/enums/ProjectAuditStatus.java
@@ -9,48 +9,48 @@ public enum ProjectAuditStatus {
     /**
      *初审通过
      */
-    AUDIT_PASS(1,"初审通过"),
+    AUDIT_PASS((byte)1,"初审通过"),
 
     /**
      * 初审打回
      */
-    AUDIT_REPULSE(2,"初审打回"),
+    AUDIT_REPULSE((byte)2,"初审打回"),
 
     /**
      * 复审通过
      */
-    RECHECK_PASS(3,"复审通过"),
+    RECHECK_PASS((byte)3,"复审通过"),
 
     /**
      * 复审拒绝
      */
-    RECHECK_REJECT(4,"复审拒绝"),
+    RECHECK_REJECT((byte)4,"复审拒绝"),
 
     /**
      * 复审打回
      */
-    RECHECK_REPULSE(5,"复审打回"),
+    RECHECK_REPULSE((byte)5,"复审打回"),
 
     /**
      * 满标复审通过
      */
-    FULL_RECHECK_PASS(6,"满标复审通过"),
+    FULL_RECHECK_PASS((byte)6,"满标复审通过"),
 
     /**
      * 产品撤回
      */
-    REVOCATION(7,"产品撤回");
+    REVOCATION((byte)7,"产品撤回");
 
-    private int code;
+    private byte code;
 
     private String name;
 
-    ProjectAuditStatus(int code, String name) {
+    ProjectAuditStatus(byte code, String name) {
         this.code = code;
         this.name = name;
     }
 
-    public int getCode() {
+    public byte getCode() {
         return code;
     }
